requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.21.3
packaging==24.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.30
typing_extensions==4.12.2
Werkzeug==3.0.3
